fix balance

+++ internal/model/model.go
@@ -21,13 +21,13 @@ type UserBalance struct {
 
 type Withdrawal struct {
 	Order       string    `json:"order"`
-	Sum         int       `json:"sum"`
+	Sum         float64   `json:"sum"`
 	ProcessedAt time.Time `json:"processed_at"`
 }
 type OrderWithdrawal struct {
 	User  *string `json:"user,omitempty"`
 	Order string  `json:"order"`
-	Sum   int     `json:"sum"`
+	Sum   float64 `json:"sum"`
 }
 
 type AccrualResponse struct {
